micromobilitykinematics.jl Library Functions

This section outlines the library's functions along with their doc strings, enabling users to understand and leverage the capabilities of each function effectively.

User level

The User Level functions offer an intuitive interface tailored for non-expert users. These functions simplify complex kinematic computations and visualizations, making the library accessible for practical applications in micromobility design, simulation, and optimization. By organizing functions into categories such as core calculations, plotting, and data export, this section helps users quickly locate the tools they need.

core functions

The core functions in the micromobilitykinematics.jl library form the foundation for kinematic calculations and updates. They include functions for computing steering and suspension kinematics in both neutral and moved states. Additionally, the library provides angleδi and angleδo functions to calculate the steering angles of the inner and outer wheels, respectively. Together, these tools offer a robust framework for analyzing and simulating mechanical systems.

micromobilitykinematics.random_search — Function

random_search(upper_border::Tuple{Float64, Float64, Float64, Float64},lower_border::Tuple{Float64, Float64, Float64, Float64},max_angleConfig; info = false, radius = 3500, step_size = 1 )

random search with given border for the parameters and given angular area for rotary component

Arguments:

upper_border::Tuple{Float64, Float64, Float64, Float64}: upper border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (100.0, 140.0,150.0,270.0))

lower_border::Tuple{Float64, Float64, Float64, Float64}: lower border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (50.0,100.0, 100.0, 100.0))

max_angleConfig: maximal angular area for rotary component (defult: (0,35))

Keywords:

info::Bool: true if info should be printed

radius: desired track circle radius (defult: 3500)

step_size: step_size in which the angular area should be checked

Returns:

compLength: tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length)

micromobilitykinematics.steeringkinematics! — Function

steeringkinematics!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

To fully describe the kinematics of the steering system, both MOVED and NEUTRAL states must be calculated

Arguments

angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated - θx: Angle of rotation of the rotation component around the x-axis - θz: Angle of rotation of the rotation component around the z-axis

steering::Steering: Instance of a specific steering

suspension::Suspension: Instance of a specific suspension

Returns:

no returns because of in place programming

micromobilitykinematics.steeringkinematics — Function

steeringkinematics(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

To fully describe the kinematics of the steering system, both MOVED and NEUTRAL states must be calculated

Arguments

angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated - θx: Angle of rotation of the rotation component around the x-axis - θz: Angle of rotation of the rotation component around the z-axis

steering::Steering: Instance of a specific steering

suspension::Suspension: Instance of a specific suspension

Returns:

steering::Steering: Instance of a specific steering

micromobilitykinematics.steeringkinematicsNEUTRAL! — Function

steeringkinematicsNEUTRAL!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

For the rotation component in its rest position with the angles (θx, θz) = (0,0), the kinematics of the steering is calculated

Arguments

angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated - θx: Angle of rotation of the rotation component around the x-axis - θz: Angle of rotation of the rotation component around the z-axis

steering::Steering: Instance of a specific steering

suspension::Suspension: Instance of a specific suspension

Returns:

no returns because of in place programming

micromobilitykinematics.steeringkinematicsMOVED! — Function

steeringkinematicsMOVED!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

For the moving rotational component with the angles (θx, θz), the kinematics of the steering is calculated

Arguments

angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated - θx: Angle of rotation of the rotation component around the x-axis - θz: Angle of rotation of the rotation component around the z-axis

steering::Steering: Instance of a specific steering

suspension::Suspension: Instance of a specific suspension

Returns:

no returns because of in place programming

micromobilitykinematics.suspensionkinematics! — Function

suspensionkinematics!(suspension::Suspension)

suspensionkinematicsNEUTRAL!(suspension::Suspension)

calculates all the movement-dependent positions of the joints in NEUTRAL- and MOVED-Postion

Arguments:

suspension::Suspension: Suspension of the micromobility vehicle

micromobilitykinematics.suspensionkinematicsNEUTRAL! — Function

suspensionkinematicsNEUTRAL!(suspension::Suspension)

calculates all the movement-dependent positions of the joints in NEUTRAL-Position

Arguemnts:

suspension::Suspension: Suspension of the micromobility vehicle

micromobilitykinematics.suspensionkinematicsMOVED! — Function

suspensionkinematicsMOVED!(suspension::Suspension)

calculates all the movement-dependent positions of the joints in MOVED-Position

Arguments:

suspension::Suspension: Suspension of the micromobility vehicle

micromobilitykinematics.update! — Function

update!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

updates the kinematics of the given steering instance on the new angles and suspension

Arguments

angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis

steering::Steering: Instance of a specific steering

suspension::Suspension: Instance of a specific suspension

Returns:

no returns because of in place programming

micromobilitykinematics.angle_δi! — Function

angle_δi!(steering::Steering)

calculates the inner steering angle of the wheel !only possible after using function kinematics!

Arguments

steering::Steering: Instance of a specific steering

Returns:

no returns because of in place programming

micromobilitykinematics.angle_δi — Function

angle_δi(steering::Steering)

calculates the inner steering angle of the wheel !only possible after using function kinematics!

Arguments

steering::Steering: Instance of a specific steering

Returns

δ: inner steering angle of the wheel

micromobilitykinematics.angle_δo! — Function

angle_δo!(steering::Steering)

calculates the outer steering angle of the wheel !only possible after using function kinematics!

Arguments

steering::Steering: Instance of a specific steering

Returns:

no returns because of in place programming

micromobilitykinematics.angle_δo — Function

angle_δo(steering::Steering)

calculates the outer steering angle of the wheel !only possible after using function kinematics!

Arguments

steering::Steering: Instance of a specific steering

Returns

δ: inner steering angle of the wheel

Optimization

The Optimization functions in the micromobilitykinematics.jl library are designed to help users optimize various parameters related to micromobility vehicle steering kinematics. These functions include tools for performing series optimizations (optimseries), grid-based optimization searches (gridoptim), and objective calculations (steering_objective, objective°). Additionally, checkConstraints and checkConstraints° ensure that optimization results satisfy necessary system constraints.

micromobilitykinematics.optim_series — Function

n_times_parallel_optim(num::Int64, θ::Tuple{T,T},upper_border::Tuple{Float64, Float64, Float64, Float64},lower_border::Tuple{Float64, Float64, Float64, Float64},max_angleConfig)

Executes an optimisation 'n' times in parallel using threads.

!(saves Dict in folder)

Arguments:

θ::Tuple{T,T}: angle pair to be considered

upper_border::Tuple{Float64, Float64, Float64, Float64}: upper border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (100.0, 140.0,150.0,270.0))

lower_border::Tuple{Float64, Float64, Float64, Float64}: lower border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (50.0,100.0, 100.0, 100.0))

max_angleConfig: maximal angular area for rotary component (default: (0,35))

Returns:

sol_dict: Dict which contains the optimal solutions.

micromobilitykinematics.optim — Function

optim(θ::Tuple{T,T},upper_border::Tuple{Float64, Float64, Float64, Float64},lower_border::Tuple{Float64, Float64, Float64, Float64},max_angleConfig)

Arguments:

θ::Tuple{T,T}: angle pair to be considered

upper_border::Tuple{Float64, Float64, Float64, Float64}: upper border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (100.0, 140.0,150.0,270.0))

lower_border::Tuple{Float64, Float64, Float64, Float64}: lower border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (50.0,100.0, 100.0, 100.0))

max_angleConfig: maximal angular area for rotary component (default: (0,35))

Keywords

param::Tuple{I,I,I,I}: If necessary, the components can be initialised individually, otherwise the values are randomised by the function that was checked for kinematic conditions.

Returns:

opda: instance of OptDa (optimization Data)

micromobilitykinematics.grid_optim — Function

grid_optim(upper_border::Tuple{Float64, Float64, Float64, Float64},lower_border::Tuple{Float64, Float64, Float64, Float64}, max_angleConfig::Tuple{I,I}) where {I::Int64}

micromobilitykinematics.checkConstraints — Function

checkConstraints(stepsize, maxangleConfig::Tuple, steering::Steering, suspension::Suspension)

checks  all constraints and dependencies

Arguments

step_size: step_size in which the angular area should be checked

maxangleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: maximal Angle of rotation of the rotation component around the x-axis -θz: maximal Angle of rotation of the rotation component around the z-axis

steering::Steering: Instance of a specific steering

suspension::Suspension: Instance of a specific suspension

Returns:

::Bool: - false: It is not possible to match the constraints in a satisfactory manner - true: It is possible to match the constraints in a satisfactory manner

micromobilitykinematics.checkConstraints° — Function

checkConstraints°(x_rotational_radius, z_rotational_radius, track_lever_length, tie_rod_length)

The wrapper function of the checkConstraints procedure is employed for the purposes of optimisation. ! function°(): symbolises that this function should only be used within the optimisation !

Arguments

x_rotational_radius: Length of the rotational component around the x-axis. (detailed info in doc)

z_rotational_radius: Length of the rotational component around the z-axis. (detailed info in doc)

track_lever_length: Length of the track lever. (detailed info in doc)

tie_rod_length: Length of tie rod. (detailed info in doc)

Returns

binary: - 0: It is not possible to match the constraints in a satisfactory manner. - 1: It is possible to match the constraints in a satisfactory manner.

micromobilitykinematics.steering_objective — Function

steering_objective(angleConfig::Tuple{T,T}, measurements::Measurements, steering::Steering, suspension::Suspension)

Calculates the distance between the optimum point of intersection of the wheel axis (normally on the rear wheel axis) and the current point of intersection of the axis.

Arguments

angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated

θx: Angle of rotation of the rotation component around the x-axis

θz: Angle of rotation of the rotation component around the z-axis

measurements::Measurements: Instance of a specific all relevant Measurements of the vehicle

steering::Steering: Instance of a specific steering

suspension::Suspension: Instance of a specific suspension

Returns

Distance between optimal and current intersection point

micromobilitykinematics.objective° — Function

objective°(θx, θz, x_rotational_radius, z_rotational_radius, track_lever_length, tie_rod_length)

The wrapper function of the steering_objective procedure is employed for the purposes of optimisation. ! function°(): symbolises that this function should only be used within the optimisation !

Arguments

angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated

θx: Angle of rotation of the rotation component around the x-axis

θz: Angle of rotation of the rotation component around the z-axis

measurements::Measurements: Instance of a specific all relevant Measurements of the vehicle

steering::Steering: Instance of a specific steering

suspension::Suspension: Instance of a specific suspension

Returns

Distance between optimal and current intersection point

Plots

The plotting functions allow users to visualize data and results related to kinematic and optimization studies. With plotoptdaseries, users can generate time-series visualizations of performance metrics. Meanwhile, plotoptdagirdδ and plotoptdagirdobj provide intuitive grid-based plots for steering angle relationships and optimization objectives, respectively. These tools facilitate deeper insights and a more intuitive understanding of complex data

micromobilitykinematics.plot_optda_series — Function

plot_optda_series(optda_dict::Dict{Int64,Any})

micromobilitykinematics.plot_optda_gird_δ — Function

plot_optda_gird_δi(θ_max::Tuple{I,I}, steering::Steering) where {T <: Any}

Create a plot in which the turning angles of the vehicle are mapped to the turning angles of the rotation component (θx, θz) of the steering system.

Arguments

-θ_max::Tuple{I,I}: maximal turning angles of the rotation component -steering::Steering: Instance of a specific steering mechanism in which the kinematics were previously calculated.

Returns

micromobilitykinematics.plot_optda_gird_obj — Function

plot_optda_gird_obj(θ_max::Tuple{I,I}, steering::Steering, suspension::Suspension, chassi::Chassis) where {I <: Any}

Create a plot in which the objective (Ackermann ratio) are mapped to the turning angles of the rotation component (θx, θz) of the steering system.

Arguments

-θ_max::Tuple{I,I}: maximal turning angles of the rotation component -steering::Steering: Instance of a specific steering mechanism in which the kinematics were previously calculated. -suspension::Suspension: Instance of a specific suspension in which the kinematics were previously calculated. -chassi::Chassis: Instance of a Chassis

Returns

Export

The export function exportXML enables users to efficiently save and share their configurations in an XML format. This functionality is particularly useful for integrating results into other software workflows or for long-term archiving, ensuring compatibility and traceability across platforms and projects.

micromobilitykinematics.exportXML — Function

exportXML(inst::Union{Steering, Suspension}; file_name="SteeringParamList", path=@__DIR__ )

All important steering/suspension kinematics parameters are transferred to an XML file.

Arguemnts

-inst::Union{Steering, Suspension}: Instance of a steering/suspension/... where the kinematic! function has already been used.

Keywords

-path=@__DIR__: Path to the folder where the xml file will be saved.