micromobilitykinematics.jl Library Functions

This section outlines the library's functions along with their doc strings, enabling users to understand and leverage the capabilities of each function effectively.

User level

The User Level functions offer an intuitive interface tailored for non-expert users. These functions simplify complex kinematic computations and visualizations, making the library accessible for practical applications in micromobility design, simulation, and optimization. By organizing functions into categories such as core calculations, plotting, and data export, this section helps users quickly locate the tools they need.

core functions

The core functions in the micromobilitykinematics.jl library form the foundation for kinematic calculations and updates. They include functions for computing steering and suspension kinematics in both neutral and moved states. Additionally, the library provides angleδi and angleδo functions to calculate the steering angles of the inner and outer wheels, respectively. Together, these tools offer a robust framework for analyzing and simulating mechanical systems.

micromobilitykinematics.random_search — Function

random_search(upper_border::Tuple{Float64, Float64, Float64, Float64},lower_border::Tuple{Float64, Float64, Float64, Float64},max_angleConfig; info = false, radius = 3500, step_size = 1 )

random search with given border for the parameters and given angular area for rotary component

#Arguments: -upper_border::Tuple{Float64, Float64, Float64, Float64}: upper border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (100.0, 140.0,150.0,270.0)) -lower_border::Tuple{Float64, Float64, Float64, Float64}: lower border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (50.0,100.0, 100.0, 100.0)) -max_angleConfig: maximal angular area for rotary component (defult: (0,35))

#Keywords:

-info::Bool: true if info should be printed -radius: desired track circle radius (defult: 3500) -step_size: step_size in which the angular area should be checked

#Returns: -compLength: tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length)

micromobilitykinematics.steeringkinematics! — Function

steeringkinematics!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

To fully describe the kinematics of the steering system, both MOVED and NEUTRAL states must be calculated

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns:

no returns because of in place programming

micromobilitykinematics.steeringkinematics — Function

steeringkinematics(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

To fully describe the kinematics of the steering system, both MOVED and NEUTRAL states must be calculated

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns:

"steering::Steering": Instance of a steering

micromobilitykinematics.steeringkinematicsNEUTRAL! — Function

steeringkinematicsNEUTRAL!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

For the rotation component in its rest position with the angles (θx, θz) = (0,0), the kinematics of the steering is calculated

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns:

no returns because of in place programming

micromobilitykinematics.steeringkinematicsMOVED! — Function

steeringkinematicsMOVED!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

For the moving rotational component with the angles (θx, θz), the kinematics of the steering is calculated

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns:

no returns because of in place programming

micromobilitykinematics.suspensionkinematics! — Function

suspensionkinematics!(suspension::Suspension)

suspensionkinematicsNEUTRAL!(suspension::Suspension)

calculates all the movement-dependent positions of the joints in NEUTRAL- and MOVED-Postion

#Arguments: -suspension::Suspension: Suspension of the micromobility vehicle

micromobilitykinematics.suspensionkinematicsNEUTRAL! — Function

suspensionkinematicsNEUTRAL!(suspension::Suspension)

calculates all the movement-dependent positions of the joints in NEUTRAL-Position

#Arguemnts: -suspension::Suspension: Suspension of the micromobility vehicle

micromobilitykinematics.suspensionkinematicsMOVED! — Function

suspensionkinematicsMOVED!(suspension::Suspension)

calculates all the movement-dependent positions of the joints in MOVED-Position

#Arguments: -suspension::Suspension: Suspension of the micromobility vehicle

micromobilitykinematics.update! — Function

update!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

updates the kinematics of the given steering instance on the new angles and suspension

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns:

no returns because of in place programming

micromobilitykinematics.angle_δi! — Function

angle_δi!(steering::Steering)

calculates the inner steering angle of the wheel !only possible after using function kinematics!

#Arguments -steering::Steering: Instance of a specific steering

#Returns:

no returns because of in place programming

micromobilitykinematics.angle_δi — Function

angle_δi(steering::Steering)

calculates the inner steering angle of the wheel !only possible after using function kinematics!

#Arguments -steering::Steering: Instance of a specific steering

#Returns -δ: inner steering angle of the wheel

micromobilitykinematics.angle_δo! — Function

angle_δo!(steering::Steering)

calculates the outer steering angle of the wheel !only possible after using function kinematics!

#Arguments -steering::Steering: Instance of a specific steering

#Returns:

no returns because of in place programming

micromobilitykinematics.angle_δo — Function

angle_δo(steering::Steering)

calculates the outer steering angle of the wheel !only possible after using function kinematics!

#Arguments -steering::Steering: Instance of a specific steering

#Returns -δ: inner steering angle of the wheel

Optimization

The Optimization functions in the micromobilitykinematics.jl library are designed to help users optimize various parameters related to micromobility vehicle steering kinematics. These functions include tools for performing series optimizations (optimseries), grid-based optimization searches (gridoptim), and objective calculations (steering_objective, objective°). Additionally, checkConstraints and checkConstraints° ensure that optimization results satisfy necessary system constraints.

micromobilitykinematics.optim_series — Function

n_times_parallel_optim(num::Int64, θ::Tuple{T,T},upper_border::Tuple{Float64, Float64, Float64, Float64},lower_border::Tuple{Float64, Float64, Float64, Float64},max_angleConfig)

Executes an optimisation 'n' times in parallel using threads.

!(saves Dict in folder)

#Arguments: -θ::Tuple{T,T}: angle pair to be considered -upper_border::Tuple{Float64, Float64, Float64, Float64}: upper border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (100.0, 140.0,150.0,270.0)) -lower_border::Tuple{Float64, Float64, Float64, Float64}: lower border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (50.0,100.0, 100.0, 100.0)) -max_angleConfig: maximal angular area for rotary component (default: (0,35))

#Returns: -sol_dict: Dict which contains the optimal solutions.

micromobilitykinematics.optim — Function

optim(θ::Tuple{T,T},upper_border::Tuple{Float64, Float64, Float64, Float64},lower_border::Tuple{Float64, Float64, Float64, Float64},max_angleConfig)

#Arguments: -θ::Tuple{T,T}: angle pair to be considered -upper_border::Tuple{Float64, Float64, Float64, Float64}: upper border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (100.0, 140.0,150.0,270.0)) -lower_border::Tuple{Float64, Float64, Float64, Float64}: lower border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (50.0,100.0, 100.0, 100.0)) -max_angleConfig: maximal angular area for rotary component (default: (0,35))

#Keywords -param::Tuple{I,I,I,I}: If necessary, the components can be initialised individually, otherwise the values are randomised by the function that was checked for kinematic conditions.

#Returns: -opda: instance of OptDa (optimization Data)

micromobilitykinematics.grid_optim — Function

grid_optim(upper_border::Tuple{Float64, Float64, Float64, Float64},lower_border::Tuple{Float64, Float64, Float64, Float64}, max_angleConfig::Tuple{I,I}) where {I::Int64}

micromobilitykinematics.checkConstraints — Function

checkConstraints(stepsize, maxangleConfig::Tuple, steering::Steering, suspension::Suspension)

checks  all constraints and dependencies

#Arguments -step_size: step_size in which the angular area should be checked -maxangleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: maximal Angle of rotation of the rotation component around the x-axis -θz: maximal Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns: -::Bool: -false: It is not possible to match the constraints in a satisfactory manner -true: It is possible to match the constraints in a satisfactory manner

micromobilitykinematics.checkConstraints° — Function

checkConstraints°(x_rotational_radius, z_rotational_radius, track_lever_length, tie_rod_length)

The wrapper function of the checkConstraints procedure is employed for the purposes of optimisation. ! function°(): symbolises that this function should only be used within the optimisation !

#Arguments -x_rotational_radius: Length of the rotational component around the x-axis. (detailed info in doc) -z_rotational_radius: Length of the rotational component around the z-axis. (detailed info in doc) -track_lever_length: Length of the track lever. (detailed info in doc) -tie_rod_length: Length of tie rod. (detailed info in doc)

#Returns -binary: -0: It is not possible to match the constraints in a satisfactory manner. -1: It is possible to match the constraints in a satisfactory manner.

micromobilitykinematics.steering_objective — Function

steering_objective(angleConfig::Tuple{T,T}, measurements::Measurements, steering::Steering, suspension::Suspension)

Calculates the distance between the optimum point of intersection of the wheel axis (normally on the rear wheel axis) and the current point of intersection of the axis.

#Arguments
-`angleConfig::Tuple{T,T}`: angles (θx,θz) in which the rotational component is rotated
    -`θx`: Angle of rotation of the rotation component around the x-axis
    -`θz`: Angle of rotation of the rotation component around the z-axis
-`measurements::Measurements`: Instance of a specific all relevant Measurements of the vehicle
-`steering::Steering`: Instance of a specific steering
-`suspension::Suspension`: Instance of a specific suspension

#Returns
- Distance between optimal and current intersection point

micromobilitykinematics.objective° — Function

objective°(θx, θz, x_rotational_radius, z_rotational_radius, track_lever_length, tie_rod_length)

The wrapper function of the steering_objective procedure is employed for the purposes of optimisation. ! function°(): symbolises that this function should only be used within the optimisation !

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -measurements::Measurements: Instance of a specific all relevant Measurements of the vehicle -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns

Distance between optimal and current intersection point

Plots

The plotting functions allow users to visualize data and results related to kinematic and optimization studies. With plotoptdaseries, users can generate time-series visualizations of performance metrics. Meanwhile, plotoptdagirdδ and plotoptdagirdobj provide intuitive grid-based plots for steering angle relationships and optimization objectives, respectively. These tools facilitate deeper insights and a more intuitive understanding of complex data

micromobilitykinematics.plot_optda_series — Function

plot_optda_series(optda_dict::Dict{Int64,Any})

micromobilitykinematics.plot_optda_gird_δ — Function

plot_optda_gird_δi(θ_max::Tuple{I,I}, steering::Steering) where {T <: Any}

Create a plot in which the turning angles of the vehicle are mapped to the turning angles of the rotation component (θx, θz) of the steering system.

#Arguments -θ_max::Tuple{I,I}: maximal turning angles of the rotation component -steering::Steering: Instance of a specific steering mechanism in which the kinematics were previously calculated.

#Returns

micromobilitykinematics.plot_optda_gird_obj — Function

plot_optda_gird_obj(θ_max::Tuple{I,I}, steering::Steering, suspension::Suspension, chassi::Chassis) where {I <: Any}

Create a plot in which the objective (Ackermann ratio) are mapped to the turning angles of the rotation component (θx, θz) of the steering system.

#Arguments -θ_max::Tuple{I,I}: maximal turning angles of the rotation component -steering::Steering: Instance of a specific steering mechanism in which the kinematics were previously calculated. -suspension::Suspension: Instance of a specific suspension in which the kinematics were previously calculated. -chassi::Chassis: Instance of a Chassis

#Returns

Export

The export function exportXML enables users to efficiently save and share their configurations in an XML format. This functionality is particularly useful for integrating results into other software workflows or for long-term archiving, ensuring compatibility and traceability across platforms and projects.

micromobilitykinematics.exportXML — Function

exportXML(inst::Union{Steering, Suspension}; file_name="SteeringParamList", path=@__DIR__ )

All important steering/suspension kinematics parameters are transferred to an XML file.

#Arguemnts -inst::Union{Steering, Suspension}: Instance of a steering/suspension/... where the kinematic! function has already been used.

#Keywords -path=@__DIR__: Path to the folder where the xml file will be saved.

All Functions

micromobilitykinematics.AngleDependency — Method

AngleDependency(steering::Steering)

checks angle dependency

#Arguments

steering::Steering: Instance of a specific steering

#Returns: -::Bool

micromobilitykinematics.EulerRotationAngles — Method

EulerRotationAngles(u1,v1,w1,u2,v2,w2,sequence)

calculates the Euler angles of rotation

Arguments

u1: base vector Coordinate System 1

v1: base vector Coordinate System 1

w1: base vector Coordinate System 1

u2: base vector Coordinate System 2

v2: base vector Coordinate System 2

w2: base vector Coordinate System 2

sequence: vector as a sequence of (x,y,z) to be rotated from System 1 to System 2

Returns

alpha, beta, gamma: Rotation angles

micromobilitykinematics.KinematicDependency — Method

KinematicDependency(steering::Steering)

checks kinematic dependency

#Arguments -steering::Steering: Instance of a specific steering

#Returns: -::Bool

micromobilitykinematics.SingularityConstraint — Method

SingularityConstraint(steering::Steering, suspension::Suspension)

checks singularity constraints

#Arguments

steering::Steering: Instance of a specific steering

suspension::Suspension: Instance of a specific suspension

#Returns: -::Bool

micromobilitykinematics.TrackingCircleConstraint — Method

TrackingcircleConstraint(steering::Steering, measurements::Measurements)

checks tracking circle constraint

#Arguments -steering::Steering: Instance of a specific steering -measurements::Measurements: Instance of a specific all relevant Measurements of the vehicle

#Returns: -::Bool

micromobilitykinematics.angle_dependency — Method

angle_dependency(steering_now::Steering)

calculates the Difference of the front wheels steering angles. !The inner steering angle should always be bigger then the outer!

!Instead of calling the kinematics in all conditions, the approach taken here is to use the properties of instantiation! !marked with a at the beginning ! #Arguments

steering_now::Steering: the current state of the steering kinematics

#Returns

Difference between the current turning angles of the front wheels

micromobilitykinematics.angle_δi! — Method

angle_δi!(steering::Steering)

calculates the inner steering angle of the wheel !only possible after using function kinematics!

#Arguments -steering::Steering: Instance of a specific steering

#Returns:

no returns because of in place programming

micromobilitykinematics.angle_δi — Method

angle_δi(steering::Steering)

calculates the inner steering angle of the wheel !only possible after using function kinematics!

#Arguments -steering::Steering: Instance of a specific steering

#Returns -δ: inner steering angle of the wheel

micromobilitykinematics.angle_δo! — Method

angle_δo!(steering::Steering)

calculates the outer steering angle of the wheel !only possible after using function kinematics!

#Arguments -steering::Steering: Instance of a specific steering

#Returns:

no returns because of in place programming

micromobilitykinematics.angle_δo — Method

angle_δo(steering::Steering)

calculates the outer steering angle of the wheel !only possible after using function kinematics!

#Arguments -steering::Steering: Instance of a specific steering

#Returns -δ: inner steering angle of the wheel

micromobilitykinematics.applyMatrixRotation — Method

applyMatrixRotation(vec::SVector{3,T}, basis1::Matrix{T}, basis2::Matrix{T}) where {T}

performs a rotation of a vector vec using two sets of three orthogonal unit vectors basis1 and basis2.

micromobilitykinematics.checkConstraints — Method

checkConstraints(stepsize, maxangleConfig::Tuple, steering::Steering, suspension::Suspension)

checks  all constraints and dependencies

#Arguments -step_size: step_size in which the angular area should be checked -maxangleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: maximal Angle of rotation of the rotation component around the x-axis -θz: maximal Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns: -::Bool: -false: It is not possible to match the constraints in a satisfactory manner -true: It is possible to match the constraints in a satisfactory manner

micromobilitykinematics.checkConstraints° — Method

checkConstraints°(x_rotational_radius, z_rotational_radius, track_lever_length, tie_rod_length)

The wrapper function of the checkConstraints procedure is employed for the purposes of optimisation. ! function°(): symbolises that this function should only be used within the optimisation !

#Arguments -x_rotational_radius: Length of the rotational component around the x-axis. (detailed info in doc) -z_rotational_radius: Length of the rotational component around the z-axis. (detailed info in doc) -track_lever_length: Length of the track lever. (detailed info in doc) -tie_rod_length: Length of tie rod. (detailed info in doc)

#Returns -binary: -0: It is not possible to match the constraints in a satisfactory manner. -1: It is possible to match the constraints in a satisfactory manner.

micromobilitykinematics.copy — Method

copy(steering::Steering)

creates a copy of the steering instance

#Arguments: -steering::Steering: instance of a Steering

#Returns: -copy::Steering: copy of the given instance

micromobilitykinematics.create_model — Method

createspheremodel(θ::Tuple{T,T})

creates model for optimization with sphere dependencies

#Arguments -θ::Tuple{T,T}: tuple of Angles (θx, θz) -start_set: Parameter start values ["xrotationalradius";"zrotationalradius";"trackleverlength";"tierodlength"]

#Returns: -model: returns a jump model

See Also

micromobilitykinematics.exportXML — Method

exportXML(inst::Union{Steering, Suspension}; file_name="SteeringParamList", path=@__DIR__ )

All important steering/suspension kinematics parameters are transferred to an XML file.

#Arguemnts -inst::Union{Steering, Suspension}: Instance of a steering/suspension/... where the kinematic! function has already been used.

#Keywords -path=@__DIR__: Path to the folder where the xml file will be saved.

micromobilitykinematics.getValue — Method

getValue(steering::Steering)

Returns important length values of steering components

#Arguments -steering::Steering: Instance of a specific steering

#Returns -(x_rotational_radius, z_rotational_radius, track_lever_length, tie_rod_length): important length values of steering components

micromobilitykinematics.get_insights — Method

get_insights(steering::Steering)

micromobilitykinematics.get_model_solution — Method

get_model_solution(model)

gets the importend informations of the model solution

#Arguments: -model: jump model

#Returns: -objectiv: value of the final objective

micromobilitykinematics.grid_data_obj — Method

grid_data_obj(θ_max::Tuple{I,I}, steering::Steering, suspension::Suspension, chassi::Chassis; step_size = 1) where {I <: Any}

calculates the objective of the given

micromobilitykinematics.grid_data_δ — Method

grid_data_δ(θ_max::Tuple{I,I}, steering::Steering; step_size = 1) where {I <: Any}

Calculates the turning angles of the vehicle for each turning angles of the rotation component (θx, θz) of the steering system.

#Arguments -θ_max::Tuple{I,I}: maximal turning angles of the rotation component -steering::Steering: Instance of a specific steering mechanism in which the kinematics were previously calculated.

#Keywords -step_size: The step size of the grid

#Retruns -δi, δo: Tuple of turning angles Matrix

micromobilitykinematics.grid_optim — Method

grid_optim(upper_border::Tuple{Float64, Float64, Float64, Float64},lower_border::Tuple{Float64, Float64, Float64, Float64}, max_angleConfig::Tuple{I,I}) where {I::Int64}

micromobilitykinematics.handle_instance — Method

handle_instance(child::XMLElement, inst::Union{T,D}) where {T<:AbstractSteering, D <:AbstractSuspension}

The inheritance tree of the instance is followed until each parameter of the instance is included in the XML file, thus ensuring that all relevant information is conveyed within the file.

#Arguments -child::XMLElement: Child to which the properties should be added in XML. -inst::Union{<:AbstractSteering,<:AbstractSuspension}: instance that is viewed

micromobilitykinematics.handle_single — Method

handle_single(child::XMLElement, field::Symbol, inst::Union{T,D}) where {T<:AbstractSteering, D <:AbstractSuspension}

Single values that exist in the core.jl module are transferred to the XML file.

#Arguments -child::XMLElement: Child to which the properties should be added in XML. -field::Symbol: the current parameter of the instance being discussed -inst::Union{<:AbstractSteering,<:AbstractSuspension}: instance that is viewed

micromobilitykinematics.handle_tuple — Method

handle_tuple(child::XMLElement, field::Symbol, inst::Union{T,D}) where {T<:AbstractSteering, D <:AbstractSuspension}

Tuples that exist in the core.jl module are transferred to the XML file.

#Arguments -child::XMLElement: Child to which the properties should be added in XML. -field::Symbol: the current parameter of the instance being discussed -inst::Union{<:AbstractSteering,<:AbstractSuspension}: instance that is viewed

micromobilitykinematics.handle_vector — Method

handle_vector(child::XMLElement, field::Symbol, inst::Union{T,D}) where {T<:AbstractSteering, D <:AbstractSuspension}

Vectors that exist in the core.jl module are transferred to the XML file.

#Arguments -child::XMLElement: Child to which the properties should be added in XML. -field::Symbol: the current parameter of the instance being discussed -inst::Union{<:AbstractSteering,<:AbstractSuspension}: instance that is viewed

micromobilitykinematics.inner_singularity_constraint — Method

inner_singularity_constraint(steering_now::Steering, steering_next::Steering)

calculates the angle change of two sequential steps of the right front wheel. The result of the subtraction should always be negativ.

!Instead of calling the kinematics in all conditions, the approach taken here is to use the properties of instantiation! !marked with a at the beginning ! #Arguments

steering_now::Steering: the current state of the steering kinematics

steering_next::Steering: The state of the steering kinematics when the rotating part of the steering is rotated about θz+1.

#Returns:

Difference between the current inner turning angle of the wheel and the next step

micromobilitykinematics.kinematicsASOFmountMOVED°! — Method

kinematicsASOFmount°!(steering::Steering)

For the moving rotational component with angles (θx, θz) and the neutral position of the rotational component, the steering kinematics are calculated as of track_lever_mounting_points_ucs. ! function°(): symbolises that this function should only be used within the optimisation ! ! function kinematicsUNTILmountMOVED°! should already be used on the steering instance !

#Arguments -steering::Steering: Instance of a specific steering where the function was previously called

#Returns:

No return value due to in-place computation.

micromobilitykinematics.kinematicsASOFmountNEUTRAL°! — Method

kinematicsASOFmountNEUTRAL°!(steering::Steering)

for the moving rotational component with the angles (θx, θz), the kinematics of the steering is calculated as of track_lever_mounting_points_ucs ! function°(): symbolises that this function should only be used within the optimisation ! ! function kinematicsUNTILmountMOVED°! should already be used on the steering instance !

#Arguments -steering::Steering: Instance of a specific steering where the function kinematicsUNTILmount°! was previously called

#Returns:

No return value due to in-place computation.

micromobilitykinematics.kinematicsASOFmount°! — Method

kinematicsASOFmount°!(steering::Steering)

For the moving rotational component with angles (θx, θz) and the neutral position of the rotational component, the steering kinematics are calculated from track_lever_mounting_points_ucs. ! function°(): symbolises that this function should only be used within the optimisation ! ! function kinematicsUNTILmountMOVED°! should already be used on the steering instance !

#Arguments -steering::Steering: Instance of a specific steering where the function kinematicsUNTILmount°! was previously called

#Returns:

No return value due to in-place computation.

micromobilitykinematics.kinematicsUNTILmountMOVED°! — Method

kinematicsUNTILmountMOVED°!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

For the moving rotational component with the angles (θx, θz), the kinematics of the steering is calculated until track_lever_mounting_points_ucs ! function°(): symbolises that this function should only be used within the optimisation !

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns:

No return value due to in-place computation.

micromobilitykinematics.kinematicsUNTILmountNEUTRAL°! — Method

kinematicsUNTILmountNEUTRAL°!(steering::Steering)

For the rotation component with the neutral position, the steering kinematics are calculated until 'tracklevermountingpointsucs'. ! function°(): symbolises that this function should only be used within the optimisation !

#Arguments -steering::Steering: Instance of a specific steering

#Returns:

No return value due to in-place computation.

micromobilitykinematics.kinematicsUNTILmount°! — Method

kinematicsUNTILmount°!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

For the moving rotation component with the angles (θx, θz) and the neutral position of the rotation component, the steering kinematics are calculated until track_lever_mounting_points_ucs. ! function°(): symbolises that this function should only be used within the optimisation !

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns:

No return value due to in-place computation.

micromobilitykinematics.kinematicsUNTILmount° — Method

kinematicsUNTILmount°(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

For the moving rotation component with the angles (θx, θz) and the neutral position of the rotation component, the steering kinematics are calculated until track_lever_mounting_points_ucs. ! function°(): symbolises that this function should only be used within the optimisation !

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns: -steering::Steering: Instance of a steering with calculated kinematics until track_lever_mounting_points_ucs.

micromobilitykinematics.left_circcirc_max_intersection_dependency — Method

left_circcirc_max_intersection_dependency(steering_now::Steering)

calculates the Difference of the radius og the circle2 (tie rod) and the total length of the distance of the circ centeres d and the radius of the circle1 (track lever). If the value of the radius of circle2 is bigger then there is no intersect posible. circle1 lies in circle2. (Left side of the kinematic steering mechanism)

!Instead of calling the kinematics in all conditions, the approach taken here is to use the properties of instantiation!

#Arguments

steering_now::Steering: the current state of the steering kinematics

#Returns

dependency for the maximal interaction between circle and circle of the kinematic

micromobilitykinematics.left_circsphere_plane_dependency — Method

left_circsphere_plane_dependency(steering_now::Steering)

calculates the diffrence between the distance of the centers and the actual radius of the actual range of motion of the tie rod. (Left side of the kinematik steering mechanism)

!Instead of calling the kinematics in all conditions, the approach taken here is to use the properties of instantiation! !marked with a at the beginning ! #Arguments

steering_now::Steering: the current state of the steering kinematics

#Returns

dependency for the interaction between circle and sphere of the kinematic

micromobilitykinematics.load_data — Function

function load_data(θ::Tuple, path = @__DIR__)

micromobilitykinematics.objective° — Method

objective°(θx, θz, x_rotational_radius, z_rotational_radius, track_lever_length, tie_rod_length)

The wrapper function of the steering_objective procedure is employed for the purposes of optimisation. ! function°(): symbolises that this function should only be used within the optimisation !

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -measurements::Measurements: Instance of a specific all relevant Measurements of the vehicle -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns

Distance between optimal and current intersection point

micromobilitykinematics.optim — Method

optim(θ::Tuple{T,T},upper_border::Tuple{Float64, Float64, Float64, Float64},lower_border::Tuple{Float64, Float64, Float64, Float64},max_angleConfig)

#Arguments: -θ::Tuple{T,T}: angle pair to be considered -upper_border::Tuple{Float64, Float64, Float64, Float64}: upper border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (100.0, 140.0,150.0,270.0)) -lower_border::Tuple{Float64, Float64, Float64, Float64}: lower border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (50.0,100.0, 100.0, 100.0)) -max_angleConfig: maximal angular area for rotary component (default: (0,35))

#Keywords -param::Tuple{I,I,I,I}: If necessary, the components can be initialised individually, otherwise the values are randomised by the function that was checked for kinematic conditions.

#Returns: -opda: instance of OptDa (optimization Data)

micromobilitykinematics.optim_series — Method

n_times_parallel_optim(num::Int64, θ::Tuple{T,T},upper_border::Tuple{Float64, Float64, Float64, Float64},lower_border::Tuple{Float64, Float64, Float64, Float64},max_angleConfig)

Executes an optimisation 'n' times in parallel using threads.

!(saves Dict in folder)

#Arguments: -θ::Tuple{T,T}: angle pair to be considered -upper_border::Tuple{Float64, Float64, Float64, Float64}: upper border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (100.0, 140.0,150.0,270.0)) -lower_border::Tuple{Float64, Float64, Float64, Float64}: lower border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (50.0,100.0, 100.0, 100.0)) -max_angleConfig: maximal angular area for rotary component (default: (0,35))

#Returns: -sol_dict: Dict which contains the optimal solutions.

micromobilitykinematics.outer_singularity_constraint — Method

outer_singularity_constraint(steering_now::Steering, steering_next::Steering)

calculates the angle change of two sequential steps of the left front wheel. The result of the subtraction should always be negativ.

!Instead of calling the kinematics in all conditions, the approach taken here is to use the properties of instantiation! !marked with a at the beginning ! #Arguments

steering_now::Steering: the current state of the steering kinematics

steering_next::Steering: The state of the steering kinematics when the rotating part of the steering is rotated about θz+1.

#Returns:

Difference between the current outer turning angle of the wheel and the next step

micromobilitykinematics.plot_optda_gird_obj — Method

plot_optda_gird_obj(θ_max::Tuple{I,I}, steering::Steering, suspension::Suspension, chassi::Chassis) where {I <: Any}

Create a plot in which the objective (Ackermann ratio) are mapped to the turning angles of the rotation component (θx, θz) of the steering system.

#Arguments -θ_max::Tuple{I,I}: maximal turning angles of the rotation component -steering::Steering: Instance of a specific steering mechanism in which the kinematics were previously calculated. -suspension::Suspension: Instance of a specific suspension in which the kinematics were previously calculated. -chassi::Chassis: Instance of a Chassis

#Returns

micromobilitykinematics.plot_optda_gird_δ — Method

plot_optda_gird_δi(θ_max::Tuple{I,I}, steering::Steering) where {T <: Any}

Create a plot in which the turning angles of the vehicle are mapped to the turning angles of the rotation component (θx, θz) of the steering system.

#Arguments -θ_max::Tuple{I,I}: maximal turning angles of the rotation component -steering::Steering: Instance of a specific steering mechanism in which the kinematics were previously calculated.

#Returns

micromobilitykinematics.plot_optda_series — Method

plot_optda_series(optda_dict::Dict{Int64,Any})

micromobilitykinematics.random_search — Method

random_search(upper_border::Tuple{Float64, Float64, Float64, Float64},lower_border::Tuple{Float64, Float64, Float64, Float64},max_angleConfig; info = false, radius = 3500, step_size = 1 )

random search with given border for the parameters and given angular area for rotary component

#Arguments: -upper_border::Tuple{Float64, Float64, Float64, Float64}: upper border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (100.0, 140.0,150.0,270.0)) -lower_border::Tuple{Float64, Float64, Float64, Float64}: lower border Tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length) (guidline = (50.0,100.0, 100.0, 100.0)) -max_angleConfig: maximal angular area for rotary component (defult: (0,35))

#Keywords:

-info::Bool: true if info should be printed -radius: desired track circle radius (defult: 3500) -step_size: step_size in which the angular area should be checked

#Returns: -compLength: tuple (xrotationalradius, zrotationalradius, tracklever.length, tierod.length)

micromobilitykinematics.right_circcirc_max_intersection_dependency — Method

left_circcirc_max_intersection_dependency(steering_now::Steering)

calculates the Difference of the radius og the circle2 (tie rod) and the total length of the distance of the circ centeres d and the radius of the circle1 (track lever). If the value of the radius of circle2 is bigger then there is no intersect posible. circle1 lies in circle2. (Right side of the kinematic steering mechanism)

!Instead of calling the kinematics in all conditions, the approach taken here is to use the properties of instantiation!

#Arguments

steering_now::Steering: the current state of the steering kinematics

#Returns

dependency for the maximal interaction between circle and circle of the kinematic

micromobilitykinematics.right_circsphere_plane_dependency — Method

right_circsphere_plane_dependency(steering_now::Steering)

calculates the diffrence between the distance of the circ/shere centers and the actual radius of the actual range of motion of the tie rod. (Right side of the kinematic steering mechanism)

!Instead of calling the kinematics in all conditions, the approach taken here is to use the properties of instantiation! !marked with a at the beginning ! #Arguments

steering_now::Steering: the current state of the steering kinematics

#Returns

dependency for the interaction between circle and sphere of the kinematic

micromobilitykinematics.rotate3 — Function

rotate3(A,n,alpha,shift=[0;0;0])

calculates the Rotation matrix RM an rotation point R.

#Arguments

A: Point to be rotated

n: vector around which to rotate

alpha: the angle to be rotated

shift=[0;0;0]: used to rotate around axes that are not going through the origin

#Returns -RM: Rotation matrix -R: rotated point

micromobilitykinematics.save_best_objective — Method

TODO

micromobilitykinematics.save_current_best_objective° — Method

save(steering::Steering, objective::T) where {T<:Number}

saves the current best objective of the optimaization

#Arguments -steering::Steering: Instance of a specific steering -objective: objective value of the current Iteration

#Returns -nothing

micromobilitykinematics.steering_objective — Method

steering_objective(angleConfig::Tuple{T,T}, measurements::Measurements, steering::Steering, suspension::Suspension)

Calculates the distance between the optimum point of intersection of the wheel axis (normally on the rear wheel axis) and the current point of intersection of the axis.

#Arguments
-`angleConfig::Tuple{T,T}`: angles (θx,θz) in which the rotational component is rotated
    -`θx`: Angle of rotation of the rotation component around the x-axis
    -`θz`: Angle of rotation of the rotation component around the z-axis
-`measurements::Measurements`: Instance of a specific all relevant Measurements of the vehicle
-`steering::Steering`: Instance of a specific steering
-`suspension::Suspension`: Instance of a specific suspension

#Returns
- Distance between optimal and current intersection point

micromobilitykinematics.steeringkinematics! — Method

steeringkinematics!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

To fully describe the kinematics of the steering system, both MOVED and NEUTRAL states must be calculated

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns:

no returns because of in place programming

micromobilitykinematics.steeringkinematics — Method

steeringkinematics(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

To fully describe the kinematics of the steering system, both MOVED and NEUTRAL states must be calculated

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns:

"steering::Steering": Instance of a steering

micromobilitykinematics.steeringkinematicsMOVED! — Method

steeringkinematicsMOVED!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

For the moving rotational component with the angles (θx, θz), the kinematics of the steering is calculated

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns:

no returns because of in place programming

micromobilitykinematics.steeringkinematicsNEUTRAL! — Method

steeringkinematicsNEUTRAL!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

For the rotation component in its rest position with the angles (θx, θz) = (0,0), the kinematics of the steering is calculated

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns:

no returns because of in place programming

micromobilitykinematics.suspensionkinematics! — Method

suspensionkinematics!(suspension::Suspension)

suspensionkinematicsNEUTRAL!(suspension::Suspension)

calculates all the movement-dependent positions of the joints in NEUTRAL- and MOVED-Postion

#Arguments: -suspension::Suspension: Suspension of the micromobility vehicle

micromobilitykinematics.suspensionkinematicsMOVED! — Method

suspensionkinematicsMOVED!(suspension::Suspension)

calculates all the movement-dependent positions of the joints in MOVED-Position

#Arguments: -suspension::Suspension: Suspension of the micromobility vehicle

micromobilitykinematics.suspensionkinematicsNEUTRAL! — Method

suspensionkinematicsNEUTRAL!(suspension::Suspension)

calculates all the movement-dependent positions of the joints in NEUTRAL-Position

#Arguemnts: -suspension::Suspension: Suspension of the micromobility vehicle

micromobilitykinematics.track_circle_dependency — Method

track_circle_dependency(steering::Steering, measurements::Measurements)

calculates the diffrence of the max. ideal outer wheel angle and the current max. outer wheel angle.

!Instead of calling the kinematics in all conditions, the approach taken here is to use the properties of instantiation!

#Arguments

steering::Steering: the last state of the steering kinematics

-measurements::Measurements: Instance of a specific all relevant Measurements of the vehicle

#Returns

dependency for the minimal track circle

micromobilitykinematics.update! — Method

update!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

updates the kinematics of the given steering instance on the new angles and suspension

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns:

no returns because of in place programming

micromobilitykinematics.update — Method

update!(angleConfig::Tuple{T,T}, steering::Steering, suspension::Suspension) where {T<:Real}

updates the kinematics of the given steering instance on the new angles and suspension

#Arguments -angleConfig::Tuple{T,T}: angles (θx,θz) in which the rotational component is rotated -θx: Angle of rotation of the rotation component around the x-axis -θz: Angle of rotation of the rotation component around the z-axis -steering::Steering: Instance of a specific steering -suspension::Suspension: Instance of a specific suspension

#Returns -steering::Steering: Instance of a specific steering

micromobilitykinematics.update°! — Method

update°!(steering::Steering)

Determines the further kinematic calculation and determines the turning angles of the wheels

! function°(): symbolises that this function should only be used within the optimisation ! ! function kinematicsUNTILmountMOVED°! should already be used on the steering instance !

#Arguments -steering::Steering: Instance of a specific steering where the function kinematicsUNTILmount°! was previously called

#Returns:

No return value due to in-place computation.